backend/requirements.txt
fastapi
uvicorn
graphbit
python-multipart
docling
Pillow
openai
python-dotenv
pandas
pyxlsb
extract-msg
tabulate
mem0ai[qdrant]
qdrant-client
azure-identity
azure-core